fix: ensure rating_count is returned as integer (Issue #8)
- Book.search(): Cast COUNT(ub.rating) to int in SQL query
- books.js route: Change rating_count from string '0' to number 0
- Ensures consistent number type for rating_count across all sources

+++ backend/src/models/Book.js
@@ -122,7 +122,7 @@ export class Book {
     const result = await query(
       `SELECT b.book_id, b.open_library_id, b.title, b.author, b.cover_url, b.first_publish_year, b.is_custom, b.created_at,
               ROUND(AVG(ub.rating)::numeric, 1) AS average_rating,
-              COUNT(ub.rating) AS rating_count
+              COUNT(ub.rating)::int AS rating_count
        FROM books b
        LEFT JOIN user_books ub ON b.book_id = ub.book_id AND ub.rating IS NOT NULL
        WHERE b.title ILIKE $1 OR b.author ILIKE $1
